+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.25.0
 
+* `Repos`:
+  * `Cache`:
+    * All cache repos now do not have `open` vals - to avoid collisions in runtime
+
 ## 0.24.7
 
 * `Versions`: